Added `unsafe_set_next_window_size_constraints(...)` (early support for size callback)
NOTE: If the callback gets garbage-collected before the next `begin(...)`, Python will crash -- the `begin(...)` attempts to call a function object that's no longer there.

For example, this can happen when doing `unsafe_set_next_window_size_constraints(..., callback=lambda pos, size, desired: ...)`. The lambda isn't bound to a name, so the caller doesn't keep a reference to it, and python is free to garbage-collect it.